+++ src/stdlib/stdlib_color.h
@@ -1,0 +1,8 @@
+#ifndef STDLIB_COLOR_H
+#define STDLIB_COLOR_H
+
+#include "../core/interpreter.h"
+
+void stdlib_color_register(void);
+
+#endif // STDLIB_COLOR_H