+++ src/stdlib/stdlib_semver.h
@@ -1,0 +1,8 @@
+#ifndef STDLIB_SEMVER_H
+#define STDLIB_SEMVER_H
+
+#include "../core/interpreter.h"
+
+void stdlib_semver_register(void);
+
+#endif // STDLIB_SEMVER_H